Also fix features for debezium kafka
